// lunadefend/js/internal-infrastructure/s3-redirect-generator/src/types.ts
/*
 * Copyright 2021 by LunaSec (owned by Refinery Labs, Inc)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */
/**
 * An individual redirect for a dead page.
 */
export interface LinkRedirect {
  from: string;
  to: string;
}

/**
 * This is the input config that we'll read from and use to convert into the S3-specific syntax.
 */
export interface S3RedirectConfig {
  links: LinkRedirect[];
  prefix?: string;
}